use rollup CLI for build
