+++ package.json
@@ -14,8 +14,9 @@
     "test-coverage": "rm -rf coverage/* && istanbul cover --report json node_modules/.bin/_mocha -- -u exports -R spec test/test.js",
     "posttest-coverage": "remap-istanbul -i coverage/coverage-final.json -o coverage/coverage-remapped.json -b dist && remap-istanbul -i coverage/coverage-final.json -o coverage/coverage-remapped.lcov -t lcovonly -b dist && remap-istanbul -i coverage/coverage-final.json -o coverage/coverage-remapped -t html -b dist",
     "ci": "npm run test-coverage && codecov < coverage/coverage-remapped.lcov",
-    "build": "git rev-parse HEAD > .commithash && gobble build -f dist",
-    "prepublish": "npm test",
+    "build": "git rev-parse HEAD > .commithash && rollup -c",
+    "build:browser": "git rev-parse HEAD > .commithash && rollup -c rollup.config.browser.js",
+    "prepublish": "npm test && npm run build:browser",
     "lint": "eslint src"
   },
   "repository": {
@@ -46,11 +47,6 @@
     "es6-promise": "^3.0.2",
     "eslint": "^1.7.1",
     "estree-walker": "^0.2.0",
-    "gobble": "^0.10.1",
-    "gobble-babel": "^5.5.8",
-    "gobble-cli": "^0.6.0",
-    "gobble-rollup": "^0.10.0",
-    "gobble-rollup-babel": "^0.6.1",
     "istanbul": "^0.4.0",
     "magic-string": "^0.8.0",
     "mocha": "^2.3.3",
